Create exercise-34-Edabit.py
Edabit:Tic Tac Toe,

Given a tic-tac-toe board, create a function that determines whether X won, O won, or there's a tie.
The board is represented as a 2-dimensional list. A board does not have to be completely filled. Blank squares are represented with the letter "B". For each board, X begins and O goes second.